nevermind, let's not rename Switch to Toggle.
