+++ changes/4667.feature.md
@@ -1,1 +1,1 @@
-Switch has been renamed to Toggle, and can now be made to appear as either a checkbox or a switch on supported backends.
+Switch can now be made to appear as either a checkbox or a switch on supported backends.
