+++ cocoa/pyproject.toml
@@ -97,7 +97,7 @@ ScrollContainer = "toga_cocoa.widgets.scrollcontainer:ScrollContainer"
 Selection = "toga_cocoa.widgets.selection:Selection"
 Slider = "toga_cocoa.widgets.slider:Slider"
 SplitContainer = "toga_cocoa.widgets.splitcontainer:SplitContainer"
-Toggle = "toga_cocoa.widgets.toggle:Toggle"
+Switch = "toga_cocoa.widgets.switch:Switch"
 Table = "toga_cocoa.widgets.table:Table"
 TextInput = "toga_cocoa.widgets.textinput:TextInput"
 TimeInput = "toga_cocoa.widgets.timeinput:TimeInput"
